Exposed more graph-like query with new last git commit info Some refactoring
