+++ graphql/schema.graphql
@@ -12,10 +12,31 @@ type FileInfo {
   size: Int
 }
 
+type Author {
+  name: String!
+  email: String!
+  timestamp: Int!
+}
+
+type CommitDescription {
+  message: String!
+  author: Author!
+  oid: String!
+}
+
+type RepositoryV3 {
+  lastCommitDescription: CommitDescription
+  fullpath: String!
+  repoName: String!
+  id: String!
+  indexDone: Boolean!
+}
+
 type Query {
   dir(repoId: String!, path: String!): [FileInfo]!
   file(repoId: String!, path: String!): String!
   listTree(repoId: String!): [String]!
   listRepos: [Repository!]!
   refreshIndex(repoId: String!): Boolean
+  repository(repoId: String): RepositoryV3!
 }
