+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "explorook",
-  "version": "0.3.16",
+  "version": "0.3.17",
   "description": "Rookout's site addon to support local files and folders",
   "main": "dist/index.js",
   "scripts": {
@@ -12,7 +12,8 @@
     "mac-publish": "export ELECTRON_BUILDER_ALLOW_UNRESOLVED_DEPENDENCIES=true && build --mac --win --linux -p always",
     "watch": "tsc -w",
     "lint": "tslint -c tslint.json -p tsconfig.json",
-    "start": "npm run build && cross-env development=1 electron ./dist/index.js"
+    "start": "npm run build && cross-env development=1 electron ./dist/index.js",
+    "debug": "npm run build && cross-env development=1 electron --inspect-brk ./dist/index.js"
   },
   "build": {
     "productName": "Explorook",
