+++ src/server.ts
@@ -1,6 +1,7 @@
 import { GraphQLServer } from "graphql-yoga";
 import { join } from "path";
 import { repoMiddleware, resolvers, traversalMiddleware } from "./api";
+import { Request, Response, NextFunction } from "express";
 
 export const start = (accessToken: string, port: number = 44512) => {
   const typeDefs = join(__dirname, `../graphql/schema.graphql`);
@@ -11,7 +12,7 @@ export const start = (accessToken: string, port: number = 44512) => {
     middlewares: [traversalMiddleware, repoMiddleware],
   });
 
-  server.express.use((req, res, next) => {
+  server.express.use((req: Request, res: Response, next: NextFunction) => {
     if (process.env.EXPLOROOK_NOAUTH) {
       next();
       return;
@@ -25,7 +26,7 @@ export const start = (accessToken: string, port: number = 44512) => {
   });
   try {
     // tslint:disable-next-line:no-console
-    server.start({ port: port }, (options) => console.log(`Server is running on http://localhost:${options.port}`)); 
+    server.start({ port: port }, (options: { port: number }) => console.log(`Server is running on http://localhost:${options.port}`)); 
   } catch (error) {
     // tslint:disable-next-line:no-console
     console.log("couldn't start server", error);
